Add DepartmentUserDiscrepancyReport view, link in admin.
