+++ assets/test_admin.py
@@ -6,7 +6,7 @@ from django.urls import reverse
 from mixer.backend.django import mixer
 from random import randint
 
-from assets.models import HardwareAsset , SoftwareAsset
+from assets.models import HardwareAsset, SoftwareAsset
 
 User = get_user_model()
 
@@ -45,7 +45,6 @@ class AssetsAdminTestCase(TestCase):
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
         self.assertTemplateUsed(response, 'admin/hardwareasset_import_start.html')
-
 
     def test_hardwareasset_import_confirm(self):
         """Test the HardwareAssetAdmin asset_import POST response
