+++ organisation/templates/admin/organisation/departmentuser/change_list.html
@@ -7,5 +7,8 @@
     <li>
         <a href="{% url 'admin:alesco_import' %}">Upload Alesco data</a>
     </li>
+    <li>
+        <a href="{% url 'admin:departmentuser_discrepancy_report' %}">Alesco data discrepancy</a>
+    </li>
     {{ block.super }}
 {% endblock %}
